xterm-set-title: error out if NEW_TITLE is not provided on the command line
The error message is printed by the newly added f_show_usage() function, but
the usage output emitted by that function is largely aspirational -- none of
the command line options actually work yet; the only way to see the help
message text at the moment is to invoke the program without the NEW_TITLE
parameter.

This changeset also anticipates an config/build process based on the GNU Autotools;
for now the maintainer address is just hard-coded into the script.