add file: bootstrap
This is a bash script, but it is hoped that this will not be a problem
(considering that this is intended to be run by developers). There's an
assumption that bash is available (or can be made so without too much
difficulty, similar to developer-only requirements for using the GNU
Autotools).