fixes #4 [1 of 2]: autotools-based build is now functional
There are still some things that need cleaning up, but the autotools-based
build is now functional. That is, it is complete enough to generate the
programs in the project and produce a source code distribution tarball via
'make distcheck'.

We do not yet have any man page documentation, so that aspect has not yet been
accounted for. That deficiency and others will be addressed over time.