fixes #5: xterm-set-title: do proper command line options processing
This change simply drops-in my canned bash-specific command line options
parsing junk (and in so doing, fully embraces bashisms throughout). Both
short-form and ("GNU style") long-form options are supported. Options bundling
is supported for the short form opts, so both of the following are equivalent
to enable trace debugging output:

    $ xterm-set-title -v -v -v my-new-title

    $ xterm-set-title -vvv my-new-title

The '--' pseudo option is supported, and is required for those corner cases in
which you would want to set the xterm window title to a value that looks like
a command line option (that is, it starts with a leading '-' character). For
example:

    $ xterm-set-title -- --help    # sets window title to "--help"

    $ xterm-set-title -- --        # sets window title to "--"

Of course, it is never wrong to use the '--' option to separate the new window
title from the command line options, so if used from another program it might
be a good idea to always use it when invoking 'xterm-set-title'.

    $ xterm-set-title -- whatever  # sets window title to "whatever"